% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authornames.R
\docType{data}
\name{authornames}
\alias{authornames}
\title{Author names data set}
\format{A character vector}
\usage{
authornames
}
\description{
A dataset containing 731 last names of authors
}
\keyword{datasets}
