% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedataset.R
\name{balancedataset}
\alias{balancedataset}
\title{balance a data set according to some grouping factor(s)}
\usage{
balancedataset(xdata, whattobalance, n = NULL)
}
\arguments{
\item{xdata}{a \code{data.frame}}

\item{whattobalance}{a character vector with column names. The corresponding columns typically are either factor or character.}

\item{n}{integer, the number of cases to select for each factor level (or combination of factor levels)}
}
\value{
a list with 5 items
\itemize{
\item \code{$seldata} the subset of \code{xdata} with the selected rows
\item \code{$unseldata} the subset of \code{xdata} with the rows that were not selected
\item \code{$sel} the row indices of the selected rows
\item \code{$unsel} the row indices of the rows not selected
\item \code{$factors} the balance factor(s) (= \code{whattobalance})
}
}
\description{
balance a data set according to some grouping factor(s)
}
\details{
the function requires either one or two factors to be balanced over

if \code{n} is larger than the largest possible number, there will be a warning to that effect and \code{n} will be reset to the largest possible number, i.e. the function behaves as if \code{n = NULL} (the default)
}
\examples{
set.seed(1)
xdata <- data.frame(ID = sample(letters[1:4], 30, replace = TRUE),
context = sample(LETTERS[21:22], 30, replace = TRUE),
var1 = rnorm(30), var2 = rnorm(30))
table(xdata$ID, xdata$context)
balancedataset(xdata = xdata, whattobalance = c("context"), n = 2)$seldata
balancedataset(xdata = xdata, whattobalance = c("context"), n = 3)$seldata
balancedataset(xdata = xdata, whattobalance = c("context"))$seldata

# with two factors
balancedataset(xdata = xdata, whattobalance = c("context", "ID"), n = 1)$seldata

# a case where one combination occurs only once (hence n = 1): row 13 has to be in each data set
xdata2 <- xdata[-9, ]
table(xdata2$ID, xdata2$context)
x <- sapply(1:50, function(X){
  row.names(balancedataset(xdata = xdata2, whattobalance = c("context", "ID"))$seldata)
})
table(x)


}
\author{
Christof Neumann
}
