% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_model.R
\name{bootstrap_model}
\alias{bootstrap_model}
\title{refit model with bootstrapped data set}
\usage{
bootstrap_model(model, dat, N, useparallel = TRUE, allcores = FALSE)
}
\arguments{
\item{model}{a model fit from \code{lme4}}

\item{dat}{the data set for the fit (typically a \code{data.frame})}

\item{N}{the number of (re)samples to take}

\item{useparallel}{use the \code{doParallel} package?}

\item{allcores}{should all available cores be used (which might slow down any activity other than R substantially). If \code{FALSE} (default), all but one core are assigned to R.}
}
\value{
a list with model fits
}
\description{
refit model with bootstrapped data set
}
\examples{
library(lme4)
(fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy))

}
