% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheload.R
\name{cacheload}
\alias{cacheload}
\title{load knitr chunk cache interactively}
\usage{
cacheload(chunk, path = knitr::opts_chunk$get("cache.path"),
  envir = parent.frame())
}
\arguments{
\item{chunk}{character, the name of the chunk}

\item{path}{character, the path to the cache location (by default it's the one set up by knitr)}

\item{envir}{character, the environment to load the objects into (default is the parent of the function)}
}
\value{
objects are loaded and messages returned
}
\description{
load knitr chunk cache interactively
}
\details{
The function borrows heavily from the \code{qwraps2} package (via https://github.com/dewittpe/qwraps2).

The messages that display which objects, if any, were loaded and which class they belong to might not be entirely reliable.
}
\keyword{knitr}
