% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawcircle.R
\name{drawcircle}
\alias{drawcircle}
\title{add a circle to a plot}
\usage{
drawcircle(x, y, rad, npoints = 70, border = NA, ...)
}
\arguments{
\item{x}{numeric, x location}

\item{y}{numeric, y location}

\item{rad}{numeric, radius}

\item{npoints}{integer, number of points along the circumference}

\item{border}{color for circumference, by default omitted (i.e. \code{NA})}

\item{...}{other graphical parameters, e.g. \code{col}}
}
\value{
adds to current plot
}
\description{
draws a circle with vertical alignment at bottom
}
\examples{
plot(0, 0, asp=1, xlim=c(-2,2), ylim=c(-2,2), "n")
# by radius
drawcircle(-2, -2, 2, col="red")
drawcircle(-2, -2, 1, col="blue")
drawcircle(-2, -2, 0.5, col="gold")
# by area
drawcircle(2, -2, sqrt(2*2*pi/pi), col="red")
drawcircle(2, -2, sqrt(2*1*pi/pi), col="blue")
drawcircle(2, -2, sqrt(2*0.5*pi/pi), col="gold")
}
