% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementprop.R
\name{elementprop}
\alias{elementprop}
\title{Proportion of elements in a sequence}
\usage{
elementprop(S, target)
}
\arguments{
\item{S}{character of length 1 or longer, contains the sequence}

\item{target}{character of length 1: occurence of which element to model}
}
\value{
the porportion of \code{target} elements in \code{S}
}
\description{
Proportion of elements in a sequence
}
\details{
sequences cannot contain \code{NA}s at this point.
}
\examples{
elementprop("AAABBB", "B") #should be 0.5
elementprop("AAAB", "A") # should be 0.75
elementprop("AB", "A") # should be 0.5

S <- c("A", "A", "A", "B", "B", "B")
elementprop(S, target="A")
elementprop(S, target="X")
}
