% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementslope.R
\name{elementslope}
\alias{elementslope}
\title{Calculate 'slope' of elements within a sequence}
\usage{
elementslope(s, type = "linear", target = "A")
}
\arguments{
\item{s}{character of length 1 or longer, contains the sequence}

\item{type}{character (linear or logistic): which type of regression to use}

\item{target}{character of length 1: occurence of which element to model}
}
\value{
numeric of length 1
}
\description{
Calculate 'slope' of elements within a sequence
}
\details{
\code{s} can be provided via a single character, which only works if the elements are themselves only one character, for example "A" in "AAABGACCAAFW"

\code{s} can also be provided via a character vector, which will work also if the elements are themselves longer than one character, for example "Ax" in c("Ax", "B", "A", "A", "Ax", "C")
}
\examples{
S <- "AAAAABBBA"
elementslope(S, "linear", target="A")
elementslope(S, "logistic", target="A")
S <- c("A", "A", "B", "A", "A", "A")
elementslope(S, "linear")
elementslope(S, "logistic")
S <- c("Ax", "Ax", "B", "Ax", "Ax", "Ax")
elementslope(S, "linear", target = "Ax")
elementslope(S, "logistic", target = "Ax")
S <- "AAAAABBB"
elementslope(S, "linear", target="A")
elementslope(S, "logistic", target="A")
# NA because of complete separation, i.e. logistic model cannot be identified
}
