% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Shannon entropy}
\usage{
entropy(S, n = 1)
}
\arguments{
\item{S}{character of length 1 or longer, contains the sequence}

\item{n}{numeric of length 1, first- or second-order entropy (\code{n=1} or \code{n=2})}
}
\value{
a numeric value
}
\description{
Shannon entropy
}
\details{
see also the \code{\link[entropy]{entropy}} function in the \pkg{entropy}
}
\examples{
S <- "AABAACA"
entropy(S, n = 1)
\dontrun{
# entropy's version only works with table counts
St <- table(unlist(strsplit(S, "")))
entropy::entropy(St, method = "ML" , unit="log2")
}
entropy(S, n = 2)
}
