% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{expand grid with numerical component}
\usage{
expand(formula, data, res, ...)
}
\arguments{
\item{formula}{a formula}

\item{data}{the data set}

\item{res}{numerical, the resolution for the numerical variable, i.e. the length of the sequence between minimum and maximum}

\item{...}{further columns to be added with their value}
}
\value{
a data frame with columns as specified in \code{formula} and if given as in \code{\dots}
}
\description{
expand combinations with range of numerical variable
}
\details{
this function is an extension of expand.grid, i.e. it takes all the possible combinations of factors and binary variables in the data set and matches the numerical variable accordingly along a sequence from its minimum to its maximum

only combinations that actually occur in the data are returned

if \code{res = 2}, the numerical variable will be returned as range
}
\examples{
xdata <- data.frame(pred1 = rnorm(100),
                    pred2 = sample(c(0, 1), 100, T),
                    fac1 = sample(letters[1:3], 100, T),
                    fac2 = sample(LETTERS[7:8], 100, T))
expand(~pred1 + fac1 + fac2, xdata, res = 3)
expand(~pred2 + fac1 + fac2, xdata, res = 3)
adding columns
expand(~pred2 + fac1 + fac2, xdata, res = 3, x = -17, y = "gg")
}
