% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{histogram with overlay}
\usage{
histogram(num, fac, bins = 50, colours = c("red", "blue", "grey",
  "gold"), xlab = "", numrange = c(NA, NA))
}
\arguments{
\item{num}{numerical vector}

\item{fac}{vector with factor or character}

\item{bins}{numeric, the number of bins for the histogram}

\item{colours}{colours to be used for categories (by default four different colours)}

\item{xlab}{character, label for x-axis}

\item{numrange}{numeric of length 2, range for numerical values to consider}
}
\value{
a plot
}
\description{
histogram with overlay
}
\details{
\code{numrange} is by default \code{c(NA, NA)}, which does not constrain the range along the numeric variable
}
\examples{
xdata <- data.frame(var = NA, fac = sample(letters[1:3], size = 20000, replace = TRUE))
xdata$var[xdata$fac == "a"] <- rnorm(n = table(xdata$fac)["a"], mean = -2)
xdata$var[xdata$fac == "b"] <- rnorm(n = table(xdata$fac)["b"], mean = 0)
xdata$var[xdata$fac == "c"] <- rnorm(n = table(xdata$fac)["c"], mean = 1)

histogram(num = xdata$var, fac = xdata$fac, bins = 10)
histogram(num = xdata$var, fac = xdata$fac, bins = 50)
histogram(num = xdata$var, fac = xdata$fac, bins = 100)
histogram(num = xdata$var, fac = xdata$fac, bins = 50, numrange = c(-2, 2))
histogram(num = xdata$var, fac = xdata$fac, bins = 100, colours = rainbow(3))
}
