% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd.R
\name{lsd}
\alias{lsd}
\title{Levenshtein distance (or rather similarity)}
\usage{
lsd(S1, S2)
}
\arguments{
\item{S1}{character of length 1 or longer, contains the sequence}

\item{S2}{character of length 1 or longer, contains the sequence}
}
\value{
numeric
}
\description{
Levenshtein distance (or rather similarity)
}
\details{
one thing to notice is that we assume that if both sequences are uniform the mutual information is 1 (regardless of length, but in our titi data set sequence length is constant...). Also, the function will return \code{NaN} or \code{NA} if one of the sequences is uniform (exists of only one syllable/call type). Whether this is correct behaviour or a bug, I can't say at the moment...
}
