% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss.R
\name{lss}
\alias{lss}
\title{Levenshtein similarity (based on Levenshtein distance)}
\usage{
lss(S1, S2)
}
\arguments{
\item{S1}{character of length 1 or longer, contains the sequence}

\item{S2}{character of length 1 or longer, contains the sequence}
}
\value{
numeric value of Levenshtein similarity
}
\description{
Levenshtein similarity (based on Levenshtein distance)
}
\details{
we use the \code{\link[utils]{adist}} function, and normalize it according to \code{1-d(S1, S2) / max(S1l, S2l)}, where \code{D} is the \code{adist} function and \code{S1l} and \code{S2l} are the lengths of the two sequences
}
\examples{
S1 <- "AAAABBAAAC"
S2 <- "ABAABBBAAA"
lss(S1, S2)
lss(S1, S1)
\dontrun{
RecordLinkage::levenshteinSim(S1, S2)
}
S3 <- "CBX"
lss(S1, S3)
lss(S2, S3)
}
