% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf.R
\name{make_pdf}
\alias{make_pdf}
\title{compile tex files using biber for bibliographies}
\usage{
make_pdf(texfile, clean_up = FALSE, open_file = FALSE)
}
\arguments{
\item{texfile}{character, file name of tex source file}

\item{clean_up}{logical, should auxillary files be removed (by default \code{FALSE})}

\item{open_file}{logical, should the result be opened in the system viewer (by default \code{FALSE}). This probably works on MacOS anyway...}
}
\value{
an invisible list with three entries for latex console output (corresponding to three latex passes) and one item for biber console output
}
\description{
compile tex files using biber for bibliographies
}
