% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modcheck.R
\name{modcheck}
\alias{modcheck}
\title{graphical model checks}
\usage{
modcheck(mod)
}
\arguments{
\item{mod}{a model object (\code{\link[lme4]{merMod}})}
}
\value{
a plot with histogram and qq plot of model residuals and a plot with fitted versus residuals
}
\description{
graphical model checks
}
\examples{
x <- data.frame(x = rnorm(100), y = rnorm(100), ID = sample(letters, 100, replace = TRUE))
mod <- lme4::lmer(y ~ x + (1|ID), data = x)
modcheck(mod = mod)
}
