% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseq.R
\name{mseq}
\alias{mseq}
\title{make regular sequence from range of numeric vector}
\usage{
mseq(x, n = 51)
}
\arguments{
\item{x}{numeric vector}

\item{n}{length of sequence (default is 51)}
}
\value{
a numeric vector
}
\description{
make regular sequence from range of numeric vector
}
\details{
function takes a vector, calculates range (removing \code{NA}s) and returns sequence between the range limits
}
\examples{
X <- rnorm(100)
Y <- mseq(X)
}
