% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutinfo.R
\name{mutinfo}
\alias{mutinfo}
\title{Mutual information}
\source{
Matlab function by A. Kershenbaum
}
\usage{
mutinfo(S1, S2)
}
\arguments{
\item{S1}{character of length 1 or longer, contains the sequence}

\item{S2}{character of length 1 or longer, contains the sequence}
}
\value{
numeric
}
\description{
Mutual information
}
\details{
one thing to notice is that we assume that if both sequences are uniform the mutual information is 1 (regardless of length, but in our titi data set sequence length is constant...). Also, the function will return \code{NaN} or \code{NA} if one of the sequences is uniform (exists of only one syllable/call type). Whether this is correct behaviour or a bug, I can't say at the moment...
}
\examples{
S1 <- "AAABBCA"
S2 <- "AAACCAB"
mutinfo(S1, S2)

S1 <- "AAAAAAA"
S2 <- "AAAAAAA"
mutinfo(S1, S2)

S1 <- "AAAAAAB"
S2 <- "BBBBBAA"
mutinfo(S1, S2)

S1 <- "AAAAAAA"
S2 <- "BBBCCAA"
mutinfo(S1, S2)
}
