% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{ngrams}
\alias{ngrams}
\title{*n*-grams of sequences}
\usage{
ngrams(S, n, xsep = "")
}
\arguments{
\item{S}{character of length 1 or longer, contains the sequence}

\item{n}{numeric of length 1, syllable length, can only be 1, 2 or 3}

\item{xsep}{character, how should syllables be separated for 2-grams or 3-grams}
}
\value{
a data.frame with two columns, the first describes the syllables and the second is the count
}
\description{
*n*-grams of sequences
}
\details{
there is a substantially faster function in the \code{\link[tau]{textcnt}} function in the \pkg{tau} package...
}
\examples{
SS <- "AAAABBBA"
ngrams(SS, 1)
ngrams(SS, 2)
ngrams(SS, 3)
ngrams(SS, 3, xsep="-")
}
