% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbinary.R
\name{plotbinary}
\alias{plotbinary}
\title{Plot raw data for binomial data}
\usage{
plotbinary(xdata, response, pred, pstructure = NULL, agg = FALSE,
  aggcol = NULL, xlimsglobal = FALSE, cexfac = pi, bin = NULL,
  pch = NULL, col = NULL, ylim = NULL, ...)
}
\arguments{
\item{xdata}{a data frame}

\item{response}{character, the column name with the response variable (y-axis)}

\item{pred}{character, the column name with the response variable (x-axis)}

\item{pstructure}{optional, a named list the values for subsetting. Names of that list have to correspond to column names in \code{xdata}}

\item{agg}{logical, should data be first aggregated over a third variable (e.g. individual)}

\item{aggcol}{character, the column name for aggregation}

\item{xlimsglobal}{logical of length 1, or numeric of length two}

\item{cexfac}{expansion factor for points}

\item{bin}{optional, numeric of length 1 or longer, if supplied the data will be binned along 'pred' in either a specific number of bins (if length 1) or along a specific break points (of length >1)}

\item{pch}{optional point character (by default 16)}

\item{col}{optional colour (by default transparent grey)}

\item{ylim}{optional vector of length 2 for y-axis limits (by default c(0, 1))}

\item{\dots}{more arguments for the plot, for example \code{xlab=}}
}
\value{
A plot
}
\description{
Plot raw data for binomial data
}
\examples{
xdata <- data.frame(resp = sample(c(0,1), 1000, TRUE), ID = sort(sample(letters[1:20], 1000, TRUE)),
pred=sort(rnorm(1000))+rnorm(1000), cat1 = sample(1:3, 1000, TRUE), cat2 = sample(1:3, 1000, TRUE),
cat3 = sample(1:3, 1000, TRUE))
plotbinary(xdata = xdata, response = "resp", pred = "pred", agg = TRUE, aggcol = "ID")
plotbinary(xdata = xdata, response = "resp", pred = "pred", agg = TRUE, aggcol = "ID", cexfac = 1)

plotbinary(xdata = xdata, response = "resp", pred = "pred", agg = FALSE)
plotbinary(xdata = xdata, response = "resp", pred = "pred", agg = FALSE, bin=30)
}
