% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDFA.R
\name{repDFA}
\alias{repDFA}
\title{repeated DFAs with balanced subsets}
\usage{
repDFA(xdata, testfactor, balancefactor, varnames, npercomb = NULL,
  nrand = 20)
}
\arguments{
\item{xdata}{a \code{data.frame} with the data}

\item{testfactor}{character, the name of the column for the test factor (grouping factor)}

\item{balancefactor}{character of length 1 or 2, the factor(s) used for balancing}

\item{varnames}{character vector, column names of the numeric variables to be used for DFA}

\item{npercomb}{the number of cases per level of \code{balancefactor} (if one balance factor) or the number of cases per combination of \code{balancefactor} (if two balance factors)}

\item{nrand}{numeric, the number of DFAs to be calculated}
}
\value{
a \code{data.frame} with information on the first (and if appropriate also second) discriminant function
\itemize{
  \item name of the variable with the highest absolute loading
  \item loading of the variable on the given function (can be either negative or positive)
  \item proportion of variance explained by the function
}
}
\description{
can be used to determine variables that discriminate in the context of permuted DFA
}
\details{
information for second function will only be returned if there were at least 2 functions calculated
}
\examples{
data(iris)
res <- repDFA(xdata = iris, testfactor = "Species", balancefactor = "Species",
              varnames = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
              npercomb = 5, nrand = 50)
table(res$df1_best)
table(res$df2_best)
}
\references{
Berthet, M., Neumann, C., Mesbahi, G., Cäsar, C., & Zuberbühler, K. (2018). Contextual encoding in titi monkey alarm call sequences. Behavioral Ecology and Sociobiology, 72(1), 8.

Mundry, R., & Sommer, C. (2007). Discriminant function analysis with nonindependent data: consequences and an alternative. Animal Behaviour, 74(4), 965-976.
}
\seealso{
\code{\link{balancedataset}}
}
\author{
Christof Neumann
}
