% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplefunction.R
\name{resamplefunction}
\alias{resamplefunction}
\title{resample function}
\usage{
resamplefunction(model, dat, N, termsref = NULL, useparallel = TRUE,
  stepsalong = 11, allcores = FALSE)
}
\arguments{
\item{model}{a model fit from \code{glmmadmb}}

\item{dat}{the data set for the fit (typically a \code{data.frame})}

\item{N}{the number of (re)samples to take}

\item{termsref}{character with 'control' variables. These variables will be set to zero (numeric predictors) or their reference level (factors).}

\item{useparallel}{use the \code{doParallel} package?}

\item{stepsalong}{the resolution of steps along the range of numerical predictors}

\item{allcores}{should all available cores be used (which might slow down any activity other than R substantially). If \code{FALSE} (default), all but one core are assigned to R.}
}
\value{
a list with two items:
 \itemize{
  \item a \code{data.frame} with the predictor variables at the values for which the prediction was made
  \item a \code{matrix} with the same number of rows as the first element of the list, and \code{N} columns, i.e. the predicted values for each sample
 }
}
\description{
Resampling model estimates from glmmADMB fits
}
\examples{
\dontrun{
library(glmmADMB)
set.seed(12345)
N <- 100
resp <- rnorm(n = N)
pred1 <- sample(letters[1:2], length(resp), T)
pred2 <- sample(LETTERS[3:4], length(resp), T)
pred3 <- rnorm(n = N)
ref <- sample(letters[15:20], length(resp), T)
xdata <- data.frame(resp, pred1, pred2, pred3, ref)
res <- glmmadmb(resp ~ pred1*pred2 + pred3 + (1|ref), data = xdata, family = "gaussian")
summary(res)

# predict along different values of $pred3
resamplefunction(res, xdata, N = 2)
# keep $pred3 at 0
resamplefunction(res, xdata, N = 2, termsref = "pred3")
# keep $pred2 at its reference level ("C")
resamplefunction(res, xdata, N = 2, termsref = "pred2")
# keep $pred1 at reference and $pred3 at 0
resamplefunction(res, xdata, N = 2, termsref = c("pred1", "pred3"))
}

}
