% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundX.R
\name{roundX}
\alias{roundX}
\title{special rounding}
\usage{
roundX(x, roundto = 0)
}
\arguments{
\item{x}{numeric vector}

\item{roundto}{numeric of length 1, round to nearest ...}
}
\value{
numeric vector of the same length as \code{x}
}
\description{
special rounding
}
\examples{
x <- runif(20, 0, 1)
plot(roundX(x, 0.2), x)

x <- rnorm(100, 0, 1)
plot(roundX(x, 0.5), x)

}
