% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_examples.R
\name{time_examples}
\alias{time_examples}
\title{time it takes to run examples in a package}
\usage{
time_examples(package = ".", foo = NULL, includedontrun = FALSE)
}
\arguments{
\item{package}{character, path to package (default is current working
directory)}

\item{foo}{character, the name(s) of the function(s) to time. Default is set
to include all functions that are found to be exported by the
package.}

\item{includedontrun}{logical, include examples that are set to 'dontrun',
default is \code{FALSE}}
}
\value{
a data frame
}
\description{
time it takes to run examples in a package
}
\details{
This is a very experimental function. Use it with caution.
}
