% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf}
\alias{transf}
\title{Create histograms of several transformations}
\usage{
transf(vec, xround = NULL)
}
\arguments{
\item{vec}{a numerical vector}

\item{xround}{numeric, used to round \code{vec} first and the plot a line histogram}
}
\value{
A plot
}
\description{
Create histograms of several transformations
}
\examples{
x <- rnorm(1000, mean=100)
transf(x)
transf(x, 1)
x <- rpois(1000, lambda=7) + rnorm(1000)
transf(x)
transf(x, 1)
}
