% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinjitter.R
\name{violinjitter}
\alias{violinjitter}
\title{create jitter in a normal-shaped envelope}
\usage{
violinjitter(y, grouping = NULL, maxwidth = 1)
}
\arguments{
\item{y}{numeric, the response variable (along vertical axis)}

\item{grouping}{character or factor, optional, a grouping factor for \code{y}}

\item{maxwidth}{positive numeric, the maximum width along the horizontal axis}
}
\value{
a numeric vector of the same length as \code{y} that represents a relative offset to be added to the actual x-coordinates
}
\description{
create jitter in a normal-shaped envelope
}
\examples{
xdata <- data.frame(id = sample(letters[1:2], 500, TRUE))
xdata$resp[xdata$id == "a"] <- rnorm(n = sum(xdata$id == "a"), mean = -5, sd = 4)
xdata$resp[xdata$id == "b"] <- rnorm(n = sum(xdata$id == "b"), mean = 2, sd = 1)

plot(0, 0, type = "n", xlim = c(0.5, 2.5), ylim = c(-15, 7))
xdata$xcoord <- as.numeric(xdata$id)
points(xdata$xcoord, xdata$resp)

plot(0, 0, type = "n", xlim = c(0.5, 2.5), ylim = c(-15, 7))
xdata$xcoord2 <- xdata$xcoord + violinjitter(xdata$resp, grouping = xdata$id, maxwidth = 0.4)
points(xdata$xcoord2, xdata$resp, cex = 0.5)
}
