% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDFA.R
\name{repDFA}
\alias{repDFA}
\title{repeated DFAs with different subsets}
\usage{
repDFA(xdata, testfactor, balancefactor, varnames, npercomb = NULL,
  nrand = 20)
}
\arguments{
\item{xdata}{a \code{data.frame} with the data}

\item{testfactor}{character, the name of the column for the test factor}

\item{balancefactor}{character of length 1 or 2, the factors that balanced data sets should be created}

\item{varnames}{character vector, column names of the numeric variables to be used for LDA}

\item{npercomb}{the number of cases per level of \code{balancefactor} (if one balance factor) or the number of cases per combination of \code{balancefactor} (if two balance factors)}

\item{nrand}{numeric, the number of randomized LDAs to be calculated}
}
\value{
a \code{data.frame} with 7 columns (the name of variable, the loading of the function, and the percentage of variance explaine by the first (and second discriminant function))
}
\description{
repeated DFAs with different subsets
}
\details{
information for second function will only be returned if there were at least 2 functions calculated
}
\examples{
data(iris)
res <- repDFA(xdata = iris, testfactor = "Species", balancefactor = "Species",
              varnames = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
              npercomb = 5, nrand = 50)
table(res$df1_best)
table(res$df2_best)
}
