% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{transitions}
\alias{transitions}
\title{Transitions between elements of a sequence}
\usage{
transitions(S, start = TRUE, stop = TRUE, restemplate = NULL)
}
\arguments{
\item{S}{character, the sequence}

\item{start}{logical, should start state be included}

\item{stop}{logical, should end state be included}

\item{restemplate}{an optional matrix with two columns, with the transitions to look for (see details)}
}
\description{
Transitions between elements of a sequence
}
\details{
usage of \code{restemplate} might be helpful if sequences are analysed that not all contain the same/complete set of elements. This might be useful if you want to generate matching outputs even though the actual observed sequences (and hence the result of the function) might differ (e.g. in the case of letter transitions in words)
}
\examples{
S <- "AAAABB"
transitions(S)
transitions(S, start=FALSE, stop=FALSE)
}
